+++ CHANGELOG.md
@@ -13,7 +13,11 @@ All notable changes to the "emacs-mcx" extension will be documented in this file
 - Improvement of `emacs-mcx.transformTo(Upper|Lower)case` to handle the mixture of empty and non-empty selections.
 
 ## [0.19.10] - Unreleased
+### Added
 - Assign `Enter` key to close the find widget, which is like the original Emacs' behavior.
+
+### Fixed
+- Fix `kill` and `kill-whole-line` to work based on active cursors, not anchors.
 
 ## [0.19.9] - 2020-07-02
 - Fix `M-w` behavior when the find widget is visible. @youyuanwu